% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HC_Helper.R
\name{OED_Export_HiChart}
\alias{OED_Export_HiChart}
\title{OED_Export_HiChart}
\usage{
OED_Export_HiChart(hc_chart, container_id = "hc_container")
}
\value{
A formatted HTML file is generated and saved in the same folder as the article file.
}
\description{
This version of the **HiChart Helper** has been developed for testing
to recognize different chart types and extract the minimum script needed to generate a dynamic chart.
Class structures and styling will be aligned to match what is found on QualityInfo.org main.css
}
\details{
Convert a formatted HiChart table into an accessible HTML chart for use on QualityInfo.org.


**Key upgrades include:**
\itemize{
  \item Takes the hicharts object and pulls all options stored in json.
  \item QualityInfo already has hicharts.js loaded so no scripts needed in document headers.
  \item Prompt for naming the output file (no need to type \code{.html}).
  \item Saves an HTML file styled for QualityInfo.org, with CSS classes integrated for OED management.
  \item Adds skip logic and table numbering for improved screen reader accessibility.
}
}
\section{Requirements}{

\itemize{
  \item Needs rstudioapi, and htmltools packages.
  \item Users should review the Division style guide so all colors and contrasts meet requirements.
        }
}

\examples{
\dontrun{
# Launches user prompts and builds an HTML chart
#
library(StylishOED)
OED_Export_HiChart()
}
}
